requirements_test.txt
-r requirements.txt
coverage==4.3.4
mock>=1.0.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0

# Additional test requirements go here
pytest>=3.0.6
pytest-django>=3.1.2
pytest-cov>=2.4.0
pytest-pep8>=1.0.6
